split dataset - move images to different folders
Move the data images to the different folders based on what type of image it is in the dataset. There is a training, testing and validation set that has a percentage which is configurable. This only needs to be done once when the data sets need to be created, then the neural net will read the images from the different folders